stick to pageable resources for lists
