+++ subs-frontend/src/test/java/uk/ac/ebi/subs/frontend/TestConfig.java
@@ -1,0 +1,27 @@
+package uk.ac.ebi.subs.frontend;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.hateoas.hal.Jackson2HalModule;
+import org.springframework.http.MediaType;
+import org.springframework.http.converter.HttpMessageConverter;
+import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
+import org.springframework.web.client.RestTemplate;
+
+import java.util.Collections;
+
+@Configuration
+public class TestConfig {
+    @Bean
+    public RestTemplate restTemplate() {
+        final ObjectMapper mapper = new ObjectMapper();
+        mapper.registerModule(new Jackson2HalModule());
+
+        final MappingJackson2HttpMessageConverter converter = new MappingJackson2HttpMessageConverter();
+        converter.setSupportedMediaTypes(MediaType.parseMediaTypes("application/hal+json"));
+        converter.setObjectMapper(mapper);
+
+        return new RestTemplate(Collections.<HttpMessageConverter<?>>singletonList(converter));
+    }
+}
